+++ src/app/d2-vendors/Vendors.tsx
@@ -21,6 +21,7 @@ import { $rootScope } from 'ngimport';
 import { Loading } from '../dim-ui/Loading';
 import { dimVendorEngramsService } from '../vendorEngramsXyzApi/vendorEngramsXyzService';
 import { VendorDrop } from '../vendorEngramsXyzApi/vendorDrops';
+import { t } from 'i18next';
 
 interface Props {
   account: DestinyAccount;
@@ -34,6 +35,7 @@ interface State {
   ownedItemHashes?: Set<number>;
   vendorEngramDrops?: VendorDrop[];
   basePowerLevel?: number;
+  error?: Error;
 }
 
 /**
@@ -50,6 +52,10 @@ export default class Vendors extends React.Component<Props & UIViewInjectedProps
 
   // TODO: pull this into a service?
   async loadVendors() {
+    if (this.state.error) {
+      this.setState({ error: undefined });
+    }
+
     dimVendorEngramsService.getAllVendorDrops().then((vendorEngramDrops) => this.setState({ vendorEngramDrops }));
 
     // TODO: defs as a property, not state
@@ -73,12 +79,24 @@ export default class Vendors extends React.Component<Props & UIViewInjectedProps
         }
       }
     }
-    const vendorsResponse = await getVendorsApi(this.props.account, characterId);
 
-    this.setState({ defs, vendorsResponse });
+    if (!characterId) {
+      this.setState({ error: new Error("Couldn't load any characters.") });
+      return;
+    }
 
-    const trackerService = await fetchRatingsForVendors(defs, vendorsResponse);
-    this.setState({ trackerService });
+    let vendorsResponse;
+    try {
+      vendorsResponse = await getVendorsApi(this.props.account, characterId);
+      this.setState({ defs, vendorsResponse });
+    } catch (error) {
+      this.setState({ error });
+    }
+
+    if (vendorsResponse) {
+      const trackerService = await fetchRatingsForVendors(defs, vendorsResponse);
+      this.setState({ trackerService });
+    }
   }
 
   componentDidMount() {
@@ -109,8 +127,19 @@ export default class Vendors extends React.Component<Props & UIViewInjectedProps
   }
 
   render() {
-    const { defs, vendorsResponse, trackerService, ownedItemHashes, vendorEngramDrops, basePowerLevel } = this.state;
+    const { defs, vendorsResponse, trackerService, ownedItemHashes, vendorEngramDrops, basePowerLevel, error } = this.state;
     const { account } = this.props;
+
+    if (error) {
+      return (
+        <div className="vendor dim-page">
+          <div className="dim-error">
+            <h2>{t('ErrorBoundary.Title')}</h2>
+            <div>{error.message}</div>
+          </div>
+        </div>
+      );
+    }
 
     if (!vendorsResponse || !defs) {
       return <div className="vendor dim-page"><Loading/></div>;