+++ src/app/progress/Milestone.tsx
@@ -4,11 +4,14 @@ import {
 } from 'bungie-api-ts/destiny2';
 import * as React from 'react';
 import { D2ManifestDefinitions } from '../destiny2/d2-definitions.service';
-import BungieImage from '../dim-ui/BungieImage';
 import './milestone.scss';
 import RewardActivity from './RewardActivity';
 import AvailableQuest from './AvailableQuest';
 import { UISref } from '@uirouter/react';
+import Objective from './Objective';
+import { Reward } from './Reward';
+import MilestoneDisplay from './MilestoneDisplay';
+import { ActivityModifier } from './ActivityModifier';
 
 /**
  * A Milestone is an activity or event that a player can complete to earn rewards.
@@ -35,6 +38,7 @@ export function Milestone({
             milestoneDef={milestoneDef}
             availableQuest={availableQuest}
             key={availableQuest.questItemHash}
+            characterClass={character.classType}
           />
         )}
       </>
@@ -42,41 +46,59 @@ export function Milestone({
   } else if (milestone.vendors) {
     // A vendor milestone (Xur)
     return (
-      <div className="milestone-quest">
-        <div className="milestone-icon">
-          <BungieImage src={milestoneDef.displayProperties.icon} />
+      <MilestoneDisplay
+        displayProperties={milestoneDef.displayProperties}
+        description={$featureFlags.vendors
+          ? <UISref to='destiny2.vendor' params={{ id: milestone.vendors[0].vendorHash, characterId: character.characterId }}>
+              <a>{milestoneDef.displayProperties.description}</a>
+            </UISref>
+          : milestoneDef.displayProperties.description
+        }
+      />
+    );
+  } else if (milestone.activities && milestone.activities.length && milestone.rewards) {
+    // TODO: loadoutRequirementIndex
+
+    const objectives = milestone.activities[0].challenges.map((a) => a.objective);
+    if (objectives.every((objective) => objective.complete)) {
+      return null;
+    }
+
+    const modifiers = (milestone.activities[0].modifierHashes || []).map((h) => defs.ActivityModifier.get(h));
+
+    return (
+      <MilestoneDisplay
+        displayProperties={milestoneDef.displayProperties}
+      >
+        {modifiers.map((modifier) =>
+          <ActivityModifier key={modifier.hash} modifier={modifier}/>
+        )}
+        <div className="quest-objectives">
+          {milestone.activities[0].challenges.map((challenge) =>
+            <Objective defs={defs} objective={challenge.objective} key={challenge.objective.objectiveHash}/>
+          )}
         </div>
-        <div className="milestone-info">
-          <span className="milestone-name">{milestoneDef.displayProperties.name}</span>
-          <div className="milestone-description">
-            {$featureFlags.vendors
-              ? <UISref to='destiny2.vendor' params={{ id: milestone.vendors[0].vendorHash, characterId: character.characterId }}>
-                  <a>{milestoneDef.displayProperties.description}</a>
-                </UISref>
-              : milestoneDef.displayProperties.description
-            }
-          </div>
-        </div>
-      </div>
+        {milestone.rewards.map((reward) =>
+          Object.values(milestoneDef.rewards[reward.rewardCategoryHash].rewardEntries).map((entry) =>
+            entry.items.map((reward) =>
+              <Reward key={reward.itemHash} reward={reward} defs={defs}/>
+            )
+          )
+        )}
+      </MilestoneDisplay>
     );
   } else if (milestone.rewards) {
-    // Special account-wide milestones (clan)
     const rewards = milestone.rewards[0];
     const milestoneRewardDef = milestoneDef.rewards[rewards.rewardCategoryHash];
 
     return (
-      <div className="milestone-quest">
-        <div className="milestone-icon">
-          <BungieImage src={milestoneDef.displayProperties.icon} />
-        </div>
-        <div className="milestone-info">
-          <span className="milestone-name">{milestoneDef.displayProperties.name}</span>
-          <div className="milestone-description">{milestoneDef.displayProperties.description}</div>
-          {rewards.entries.map((rewardEntry) =>
-            <RewardActivity key={rewardEntry.rewardEntryHash} rewardEntry={rewardEntry} milestoneRewardDef={milestoneRewardDef} />
-          )}
-        </div>
-      </div>
+      <MilestoneDisplay
+        displayProperties={milestoneDef.displayProperties}
+      >
+        {rewards.entries.map((rewardEntry) =>
+          <RewardActivity key={rewardEntry.rewardEntryHash} rewardEntry={rewardEntry} milestoneRewardDef={milestoneRewardDef} />
+        )}
+      </MilestoneDisplay>
     );
   }