+++ src/app/progress/Objective.tsx
@@ -5,13 +5,16 @@ import classNames from 'classnames';
 import { t } from 'i18next';
 import { percent } from '../inventory/dimPercentWidth.directive';
 import BungieImage from '../dim-ui/BungieImage';
+import { settings } from '../settings/settings';
 
 export default function Objective({
   defs,
-  objective
+  objective,
+  suppressObjectiveDescription
 }: {
   defs: D2ManifestDefinitions;
   objective: DestinyObjectiveProgress;
+  suppressObjectiveDescription?: boolean;
 }) {
   const objectiveDef = defs.Objective.get(objective.objectiveHash);
 
@@ -21,8 +24,10 @@ export default function Objective({
     return null;
   }
 
-  const displayName = objectiveDef.progressDescription ||
+  const displayName = (!suppressObjectiveDescription && objectiveDef.progressDescription) ||
       t(objective.complete ? 'Objectives.Complete' : 'Objectives.Incomplete');
+
+  const formatter = new Intl.NumberFormat(settings.language);
 
   if (objectiveDef.valueStyle === DestinyUnlockValueUIStyle.Integer) {
     return (
@@ -32,7 +37,7 @@ export default function Objective({
             {objectiveDef.displayProperties.hasIcon && <BungieImage src={objectiveDef.displayProperties.icon}/>}
             {displayName}
           </div>
-          <div className="objective-text">{progress}</div>
+          <div className="objective-text">{formatter.format(progress)}</div>
         </div>
       </div>
     );
@@ -40,11 +45,11 @@ export default function Objective({
 
   const classes = classNames('objective-row', {
     'objective-complete': objective.complete,
-    'objective-boolean': objectiveDef.valueStyle === DestinyUnlockValueUIStyle.Checkbox || (objectiveDef.completionValue === 1 && !objectiveDef.allowOvercompletion)
+    'objective-boolean': objectiveDef.valueStyle === DestinyUnlockValueUIStyle.Checkbox || (objective.completionValue === 1 && !objectiveDef.allowOvercompletion)
   });
 
   const progressBarStyle = {
-    width: percent(progress / objectiveDef.completionValue)
+    width: percent(progress / objective.completionValue)
   };
 
   return (
@@ -53,9 +58,9 @@ export default function Objective({
       <div className="objective-progress">
         <div className="objective-progress-bar" style={progressBarStyle}/>
         <div className="objective-description">{displayName}</div>
-        {objectiveDef.allowOvercompletion && objectiveDef.completionValue === 1
-          ? <div className="objective-text">{progress}</div>
-          : <div className="objective-text">{progress}/{objectiveDef.completionValue}</div>
+        {objectiveDef.allowOvercompletion && objective.completionValue === 1
+          ? <div className="objective-text">{formatter.format(progress)}</div>
+          : <div className="objective-text">{formatter.format(progress)}/{formatter.format(objective.completionValue)}</div>
         }
       </div>
     </div>