+++ docs/CHANGELOG.md
@@ -1,10 +1,37 @@
 # Next
 
+* Max power updated to 600 for Forsaken owners.
+* Fixed Year 1 weapons not having an elemental damage type.
+* Many bugfixes post-Forsaken launch.
+* Add Infamy rank to progress page.
+* Bounties now show their rewards on the Progress and Vendors pages.
+* The Progress page has been cleaned up to better reflect the state of the game since Forsaken.
+* Pursuits are sorted such that bounties are displayed together.
+* Add "is:randomroll" search for items that have random rolls.
+* Added "is:bow" and "is:machinegun" searches.
+* Remove "is:powermod" and "basepower:" searches.
+* Masterworks now have a gold border. Previously items with a power mod had a gold border, but there are no more power mods.
+* Added Bow stats "Draw Time" and "Inventory Size".
+* Disabled vendorengrams.xyz integration until they are back online.
+* Review modes - say hello to Gambit (and goodbye to Trials, at least for a little while).
+* Added Etheric Spiral and Etheric Helix to the list of reputation items.
+
+# 4.68.3 (2018-09-03)
+
+# 4.68.2 (2018-09-03)
+
+# 4.68.1 (2018-09-03)
+
+# 4.68.0 (2018-09-02)
+
+* Fixed: Destiny 2 - Sort by character age.
 * Item popup shows the ammo type of D2 weapons.
 * New is:primary, is:special, and is:heavy search terms for ammo types.
 * Add is:tracerifle and is:linearfusionrifle searches.
 * Added Korean as a language option.
 * We have a new Shop selling enamel pins and T-shirts.
+* Ratings system understands random rolls in D2.
+* Search help added for searching by # of ratings.
 
 # 4.67.0 (2018-08-26)